+++ src/fourty_eight.rs
@@ -5,7 +5,7 @@ use crate::data_item::{
     CalculatedTrackVelocity, CommunicationsCapabilityFlightStatus, DataSourceIdentifier,
     FlightLevelInBinaryRepresentation, MeasuredPositionInPolarCoordinates,
     Mode3ACodeInOctalRepresentation, ModeSMBData, RadarPlotCharacteristics, TargetReportDescriptor,
-    TimeOfDay, TrackNumber, TrackQuality, TrackStatus,
+    TimeOfDay, TrackNumber, TrackQuality, TrackStatus, WarningErrorConditionsTargetClass,
 };
 use crate::fspec::{add_fx, is_fspec, read_fspec, trim_fspec};
 
@@ -57,9 +57,11 @@ pub struct Cat48 {
     pub track_status: Option<TrackStatus>,
     #[deku(skip, cond = "is_fspec(TrackQuality::FRN_48, fspec, 2)")]
     pub track_quality: Option<TrackQuality>,
-    //#[deku(skip, cond = "is_fspec(0b100_0000, fspec, 2")]
-    //pub warning_error_con_target_class = Option<WarningErrorConditionsTargetClass>,
-    //#[deku(skip, cond = "is_fspec(0b10_0000, fspec, 2")]
+    #[deku(
+        skip,
+        cond = "is_fspec(WarningErrorConditionsTargetClass::FRN_48, fspec, 2)"
+    )]
+    pub warning_error_con_target_class: Option<WarningErrorConditionsTargetClass>,
     //#[deku(skip, cond = "is_fspec(0b10_0000, fspec, 2)")]
     //pub mode3a_code_confidence_indicator = Option<Mode3ACodeConfidenceIndicator>,
     //#[deku(skip, cond = "is_fspec(0b1_0000, fspec, 2)")]
@@ -125,6 +127,9 @@ impl Cat48 {
         }
         if self.track_quality.is_some() {
             fspec[2] |= TrackQuality::FRN_48;
+        }
+        if self.warning_error_con_target_class.is_some() {
+            fspec[2] |= WarningErrorConditionsTargetClass::FRN_48;
         }
         if self.communications_capability_flight_status.is_some() {
             fspec[2] |= CommunicationsCapabilityFlightStatus::FRN_48;