+++ src/modifier.rs
@@ -2,6 +2,10 @@ pub fn groundspeed() -> f32 {
     2_f32.powi(-14)
 }
 
-pub fn heading() -> f32 {
+pub fn heading1() -> f32 {
     360.0 / 2_f32.powi(16)
+}
+
+pub fn heading2() -> f32 {
+    360.0 / 2_f32.powi(12)
 }