+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Add TrackQuality for Category 48
+- Cat48: Add TrackQuality
+- Cat48: Add WarningErrorConditionsTargetClass
 - Update Data Item docs
 
 ## [0.2.0] - 2020-08-21